Can now add images to messages
For now, a user must copy and paste an image URL into their message and
it will only render after they have already sent the message. The image
is also not resized, so it may be larger than its container. The URLS
are currently replaced by html image tags on the server when messages
are being requested